+++ types/index.d.ts
@@ -256,8 +256,8 @@ import { imageminMinify } from "./utils.js";
 import { imageminGenerate } from "./utils.js";
 import { squooshMinify } from "./utils.js";
 import { squooshGenerate } from "./utils.js";
-import { sharpMinify } from "./sharp-transformer.js";
-import { sharpGenerate } from "./sharp-transformer.js";
+import { sharpMinify } from "./utils.js";
+import { sharpGenerate } from "./utils.js";
 type Schema = import("schema-utils/declarations/validate").Schema;
 type WebpackPluginInstance = import("webpack").WebpackPluginInstance;
 type Compiler = import("webpack").Compiler;
